+++ R/helpers.R
@@ -221,27 +221,77 @@
 }
 
 # error handler
-.http_error_handler <- function(error_condition) {
+.http_error_handler <- function(error_condition, type = "ass") {
   # Use a switch or if-else to handle different types of errors
   error_type <- class(error_condition)[1]
+  # error messages
+  msg404<-switch(type,
+                 "ass" = c("x" = "Questionnaire/Assignment/Assignee not found"),
+                 "usr" = c("x" = "User not found")
+                 )
+  msg406<-switch(type,
+                 "ass" = c("x" = "Assignee cannot be assigned to assignment"),
+                 "usr" = c("x" = "User is not an interviewer or supervisor")
+  )
+
+  msg400<-switch(type,
+                 "ass" = c("x" = "Bad parameters provided or identifying data incorrect. See response details for more info"),
+                 "usr" = c("x" = "User not found")
+  )
+
+  msg401<-c("x"="Unauthorized/User not authorized.")
+
+
+  msg403<-c("x" = "Forbidden")
+
+  msg409<-switch(type,
+                 "usr" = c("x" = "User cannot be unarchived")
+                 )
+
   switch(error_type,
          "httr2_http_404" = {
-           rlang::abort("Assignment or assignee not found", parent = error_condition)
+           cli::cli_abort(
+            message = msg404,
+            call = NULL,
+            )
          },
          "httr2_http_406" = {
-           rlang::abort("Assignee cannot be assigned to assignment", parent = error_condition)
+           cli::cli_abort(
+             message = msg406,
+             call = NULL,
+           )
          },
          "httr2_http_400" = {
-           stop("User id cannot be parsed", call. = FALSE)
+           cli::cli_abort(
+             message = msg400,
+             call = NULL,
+           )
          },
          "httr2_http_401" = {
-           stop("Unauthorized", call. = FALSE)
+           cli::cli_abort(
+             message = msg401,
+             call = NULL,
+           )
          },
          "httr2_http_403" = {
-           stop("Forbidden", call. = FALSE)
+           cli::cli_abort(
+             message = msg403,
+             call = NULL,
+           )
+         },
+         "httr2_http_409" = {
+           cli::cli_abort(
+             message = msg409,
+             call = NULL,
+           )
          },
          # Default case if the error type is not handled above
-         stop("An unknown error occurred", call. = FALSE)
+         cli::cli_abort(
+           message = "The following other error occured:",
+           call = NULL,
+           parent = error_condition,
+           .internal = T
+         )
   )
 }